make BlockPoolModel aggregate
