+++ blockchain/infra/adapter/command_handler.go
@@ -10,7 +10,7 @@ var ErrBlockNil = errors.New("Block nil error");
 
 type BlockApi interface {
 	AddBlockToPool(block blockchain.Block)
-	CheckAndSaveBlockFromPool(height blockchain.BlockHeight) error
+	CheckAndSaveBlockFromPool(block blockchain.Block) error
 }
 
 type CommandHandler struct {
